implements low mach/conductive fluid flow (#22)
* weak form of low mach flows

* working on constant tests

* working on fluid

* working on some derivations

* host docu local during development

* added full q jacobians

* typos and support for w field

* added residual terms for all test functions

* change formulation for df/(du/dt) terms

* residual terms updated

* working on debug of L_2 Residual

* added gravity support

* removed pressure null space

* added mms solution file

* fixed url path

* working on residual terms

* Q-Temp

* typo in residual terms

* working jacobians?

* simple for debug

* turned back on everything

* restored null space removal

* added lincompressible flow formulation

* restored old tests

* updated test and outputs with real coeff

* added back low mach flow

* changed temp url

* cubic mms for low mach

* updated solution files with titles

* code format

* code format for c

* fixed missed formatting issues

* revert code

* debug code

* working on smart number compare

* updated tests with smart compare

* code reformat

* fixed reformat

* removed typos

+++ tests/testFixtures/MpiTestFixture.cpp
@@ -2,13 +2,18 @@
 #include <cstdlib>
 #include <filesystem>
 #include <fstream>
+#include <regex>
 
 int* MpiTestFixture::argc;
 char*** MpiTestFixture::argv;
 const std::string MpiTestFixture::InTestRunFlag = "--inmpitestrun=true";
 const std::string MpiTestFixture::Test_Mpi_Command_Name = "TEST_MPI_COMMAND";
+const std::string MpiTestFixture::Keep_Output_File = "--keepOutputFile=true";
+const std::string expectedResultDelimiter = std::string("<expects>");
 
 bool MpiTestFixture::inMpiTestRun;
+bool MpiTestFixture::keepOutputFile;
+
 std::string MpiTestFixture::mpiCommand = "mpirun";
 
 std::string MpiTestFixture::ParseCommandLineArgument(int* argc, char*** argv, const std::string flag) {
@@ -38,6 +43,7 @@ bool MpiTestFixture::InitializeTestingEnvironment(int* argc, char*** argv) {
     MpiTestFixture::argv = argv;
 
     MpiTestFixture::inMpiTestRun = !MpiTestFixture::ParseCommandLineArgument(MpiTestFixture::argc, MpiTestFixture::argv, MpiTestFixture::InTestRunFlag).empty();
+    MpiTestFixture::keepOutputFile = !MpiTestFixture::ParseCommandLineArgument(MpiTestFixture::argc, MpiTestFixture::argv, MpiTestFixture::Keep_Output_File).empty();
 
     char* commandName = std::getenv(MpiTestFixture::Test_Mpi_Command_Name.c_str());
     if (commandName != NULL) {
@@ -51,7 +57,11 @@ void MpiTestFixture::SetUp() {}
 
 void MpiTestFixture::TearDown() {
     if (!inMpiTestRun) {
-        fs::remove_all(OutputFile());
+        if (!MpiTestFixture::keepOutputFile) {
+            fs::remove_all(OutputFile());
+        } else {
+            std::cout << "Generated output file: " << OutputFile() << std::endl;
+        }
     }
 }
 
@@ -66,20 +76,90 @@ void MpiTestFixture::RunWithMPI() const {
     mpiCommand << mpiTestParameter.arguments << " ";
     mpiCommand << " > " << OutputFile();
 
-    std::system(mpiCommand.str().c_str());
+    auto exitCode = std::system(mpiCommand.str().c_str());
+    if (exitCode != 0) {
+        std::ifstream outputStream(OutputFile());
+        std::string output((std::istreambuf_iterator<char>(outputStream)), std::istreambuf_iterator<char>());
+        FAIL() << output;
+    }
 }
 
 void MpiTestFixture::CompareOutputFiles() {
-    // load the actual output
-    std::ifstream actualStream(OutputFile());
-    std::string actual((std::istreambuf_iterator<char>(actualStream)), std::istreambuf_iterator<char>());
-
-    // read in the expected
+    if (mpiTestParameter.expectedOutputFile.empty()) {
+        return;
+    }
+    // load in the actual and expected result files
     std::ifstream expectedStream(mpiTestParameter.expectedOutputFile);
-    std::string expected((std::istreambuf_iterator<char>(expectedStream)), std::istreambuf_iterator<char>());
+    std::ifstream actualStream(OutputFile());
 
-    ASSERT_TRUE(actual.length() > 0) << "Actual output is expected not to be empty";
-    ASSERT_EQ(actual, expected);
+    // march over each line
+    std::string expectedLine;
+    std::string actualLine;
+    while (std::getline(expectedStream, expectedLine)) {
+        if (!std::getline(actualStream, actualLine)) {
+            FAIL() << "The actual output file is missing lines";
+        }
+
+        // check to see if this lines includes any expected values
+        auto expectedResultDelimiterPosition = expectedLine.find(expectedResultDelimiter);
+        if (expectedResultDelimiterPosition == std::string::npos) {
+            // do a direct match
+            ASSERT_EQ(expectedLine, actualLine);
+        } else {
+            std::string regexLine = expectedLine.substr(0, expectedResultDelimiterPosition);
+            std::string valuesLine = expectedLine.substr(expectedResultDelimiterPosition + expectedResultDelimiter.size());
+
+            // get the matches
+            std::smatch matches;
+            std::regex_search(actualLine, matches, std::regex(regexLine));
+
+            // get the expected values
+            std::istringstream valuesStream(valuesLine);
+            std::vector<std::string> expectedValues(std::istream_iterator<std::string>{valuesStream}, std::istream_iterator<std::string>());
+
+            ASSERT_EQ(expectedValues.size(), matches.size() - 1) << "the number of expected and found values is different on line " << expectedLine;
+
+            // march over each value
+            for (int v = 0; v < expectedValues.size(); v++) {
+                char compareChar = expectedValues[v][0];
+                double expectedValue = stod(expectedValues[v].substr(1));
+                double actualValue = stod(matches[v + 1]);
+
+                switch (compareChar) {
+                    case '<':
+                        ASSERT_LT(actualValue, expectedValue) << " on line " << expectedLine;
+                        break;
+                    case '>':
+                        ASSERT_GT(actualValue, expectedValue) << " on line " << expectedLine;
+                        break;
+                    case '=':
+                        ASSERT_DOUBLE_EQ(actualValue, expectedValue) << " on line " << expectedLine;
+                        break;
+                    default:
+                        FAIL() << "Unknown compare char " << compareChar << " on line " << expectedLine;
+                }
+            }
+
+            //
+            //            while (valuesStream.good()) {
+            //                std::string tmp;
+            //
+            //                expectedValues.push_back(tmp);
+            //
+            //                double tmp;
+            //            }
+            //
+            //
+            //
+            //
+            //            for (auto i = 0; i < expectedValues.size(); i++) {
+            //                auto foundAsDouble = stod(matches[i + 1]);
+            //                ASSERT_DOUBLE_EQ(expectedValues[i], foundAsDouble);
+            //            }
+        }
+    }
+
+    ASSERT_FALSE(std::getline(actualStream, actualLine)) << "actual results should reach end of file";
 }
 
-std::ostream& operator<<(std::ostream& os, const MpiTestParameter& params) { return os << params.expectedOutputFile; }
+std::ostream& operator<<(std::ostream& os, const MpiTestParameter& params) { return os << (params.testName.empty() ? params.arguments : params.testName); }
